feat: Add GitHub Action workflow templates
With https://github.com/guardian/actions-assume-aws-role we're able to use GitHub Actions (GHA) for CI.

This PR introduces some workflow templates for the organisation with the aim of simplifying the migration to GHA.

There are three workflows:
  1. Node only
  2. SBT only
  3. SBT & Node - typically for use by repositories of Play apps with a React frontend, or with GuCDK infrastructure

In addition to setting up `guardian/actions-assume-aws-role`,
the workflow will also provide an ability to seed the build number from the last TeamCity number.

This is so the repository's continuous and scheduled delivery configuration in Riff-Raff still works
as Riff-Raff will only trigger these if the build number is greater than the number currently deployed.

See https://docs.github.com/en/actions/learn-github-actions/creating-workflow-templates